Split data access functions from database file
